requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
streamlit>=1.35.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
icrawler>=0.6.6
numpy>=1.24.0
